requirements.txt




huaweicloudsdkcore
huaweicloudsdkobs
huaweicloudsdkocr
huaweicloudsdkfunctiongraph
huaweicloudsdkmpc
esdk-obs-python

opencv-python
ffmpeg-python
Pillow
numpy


pytesseract


pymysql


Flask
Flask-CORS
requests


python-dotenv
colorlog
tqdm
jinja2
python-dateutil
typing-extensions
cryptography


openai